// src/models/movieModels.ts
export interface Movie {
    adult: boolean;
    backdrop_path: string | null;
    genre_ids: number[];
    id: number;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string | null;
    release_date: string;
    title: string;
    video: boolean;
    vote_average: number;
    vote_count: number;
}

export interface MovieDetail {
    adult: boolean;
    backdrop_path: string | null;
    belongs_to_collection: any | null;
    budget: number;
    genres: {
        id: number;
        name: string;
    }[];
    homepage: string;
    id: number;
    imdb_id: string;
    origin_country: string[];
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string | null;
    production_companies: {
        id: number;
        logo_path: string | null;
        name: string;
        origin_country: string;
    }[];
    production_countries: {
        iso_3166_1: string;
        name: string;
    }[];
    release_date: string;
    revenue: number;
    runtime: number;
    spoken_languages: {
        english_name: string;
        iso_639_1: string;
        name: string;
    }[];
    status: string;
    tagline: string;
    title: string;
    video: boolean;
    vote_average: number;
    vote_count: number;
}

export interface MovieCredit {
  id: number;
  cast: {
    adult: boolean;
    gender: number;
    id: number;
    known_for_department: string;
    name: string;
    original_name: string;
    popularity: number;
    profile_path: string | null;
    cast_id: number;
    character: string;
    credit_id: string;
    order: number;
  }[];
  crew: {
    adult: boolean;
    gender: number;
    id: number;
    known_for_department: string;
    name: string;
    original_name: string;
    popularity: number;
    profile_path: string | null;
    credit_id: string;
    department: string;
    job: string;
  }[];
}

export const InitialMovieCredit = {
  credit: null,
  loading: false,
  error: null
};

export const InitialMovieDetail = {
    movie: null,
    loading: false,
    error: null,
};

export interface PlayingMovies {
    dates: {
        maximum: string;
        minimum: string;
    };
    page: number;
    results: Movie[];
    total_pages: number;
    total_results: number;
}

export interface UpcomingMovies {
    dates: {
        maximum: string;
        minimum: string;
    };
    page: number;
    results: Movie[];
    total_pages: number;
    total_results: number;
}

export interface PopularMovies {
    page: number;
    results: Movie[];
    total_pages: number;
    total_results: number;
}

export interface PlayingMovieState {
    movies: Movie[];
    loading: boolean;
    page: number;
    error: string | null;
}

export interface UpcomingMovieState {
    movies: Movie[];
    loading: boolean;
    page: number;
    error: string | null;
}

export interface PopularMovieState {
    movies: Movie[];
    loading: boolean;
    page: number;
    error: string | null;
}

export const initialPlayingMovieState: PlayingMovieState = {
    movies: [],
    loading: false,
    page: 0,
    error: null,
};

export const initialUpcomingMovieState: UpcomingMovieState = {
    movies: [],
    loading: false,
    page: 0,
    error: null,
};

export const initialPopularMovieState: PopularMovieState = {
    movies: [],
    loading: false,
    page: 0,
    error: null,
};